Added rudimentary support for bitmap graphics, using PPM files and using a pallette file I called .plt which has a similar syntax to the PPM files,
I created a python script to aid me in creating both these files, I called
it VGALIZATOR and it is available in my repositories,

I say Rudimentary support because it does not work correctly, yet :),
but I already made some art intended for use in this little experiment, feel
free to use it, It is also distributed under the GPLv3 license, so go nuts.

Besides that, I separated some functions from the RENDERER.C file over to
FONFLIB.C, which is more generic in nature than the renderer file, really
playing with the idea of changing RENDERER.C to VIDEO.C, since it makes more
sense.